refactor: move buildHeadingMap to heading-map.ts + add try/catch in processFull
Addresses Copilot review feedback:
- Moved buildHeadingMap from CLI module (headingmap.ts) to shared
  module (heading-map.ts) to avoid coupling FileProcessor to CLI deps.
  Re-exported from headingmap.ts for backward compatibility.
- Wrapped heading-map injection in processFull with try/catch so a
  malformed translation (e.g. missing # title) falls back gracefully
  to returning the raw translated content with a warning.
