+++ src/file-processor.ts
@@ -22,9 +22,9 @@ import {
   updateHeadingMap, 
   lookupTargetHeading,
   injectHeadingMap,
+  buildHeadingMap,
   HeadingMap
 } from './heading-map.js';
-import { buildHeadingMap } from './cli/commands/headingmap.js';
 
 export class FileProcessor {
   private parser: MystParser;
@@ -517,21 +517,28 @@ export class FileProcessor {
 
     const translatedContent = result.translatedSection || '';
 
-    // Build heading-map from source and translated sections
-    const sourceParsed = await this.parser.parseDocumentComponents(content, filepath);
-    const translatedParsed = await this.parser.parseDocumentComponents(translatedContent, filepath);
+    // Build heading-map from source and translated sections.
+    // Wrapped in try/catch so a malformed translation (e.g. missing # title)
+    // doesn't fail the entire sync — we fall back to the raw translated content.
+    try {
+      const sourceParsed = await this.parser.parseDocumentComponents(content, filepath);
+      const translatedParsed = await this.parser.parseDocumentComponents(translatedContent, filepath);
 
-    const { map: headingMap } = buildHeadingMap(
-      sourceParsed.sections,
-      translatedParsed.sections,
-    );
+      const { map: headingMap } = buildHeadingMap(
+        sourceParsed.sections,
+        translatedParsed.sections,
+      );
 
-    const translatedTitle = translatedParsed.titleText || '';
+      const translatedTitle = translatedParsed.titleText || '';
 
-    // Inject heading-map into translated content
-    if (headingMap.size > 0 || translatedTitle) {
-      this.log(`Injecting heading-map with ${headingMap.size} entries and title "${translatedTitle}"`);
-      return injectHeadingMap(translatedContent, headingMap, translatedTitle);
+      // Inject heading-map into translated content
+      if (headingMap.size > 0 || translatedTitle) {
+        this.log(`Injecting heading-map with ${headingMap.size} entries and title "${translatedTitle}"`);
+        return injectHeadingMap(translatedContent, headingMap, translatedTitle);
+      }
+    } catch (error) {
+      const msg = error instanceof Error ? error.message : String(error);
+      core.warning(`Failed to build heading-map for ${filepath}: ${msg}. Returning translated content without heading-map.`);
     }
 
     return translatedContent;
